attack-dqn-train

+++ attack-dqn.py
@@ -22,8 +22,8 @@ parser.add_argument('-e','--env', type = str, nargs = "?", default = "PongNoFram
 parser.add_argument('-p','--perturbationType', nargs="?", default="rfgsm", type = str, help = 'Perturbation Type: fgsm, rfgsm, cw, optimal')
 parser.add_argument('-a', '--attack', nargs="?", default=1, type = int, help = 'Attack 1 or not to attack 0')
 parser.add_argument('--stepsRFGSM', nargs = "?", default = 1, type = int, help = "Number of steps of RFGSM attack")
-parser.add_argument('--alphaRFGSM', nargs = "?", default = 8/255, type = int, help = "Alpha (Step Size) of RFGSM attack")
-parser.add_argument('--epsRFGSM', nargs = "?", default = 16/255, type = int, help = "Epsilon (strength) of RFGSM attack")
+parser.add_argument('--alphaRFGSM', nargs = "?", default = 8/255, type = float, help = "Alpha (Step Size) of RFGSM attack")
+parser.add_argument('--epsRFGSM', nargs = "?", default = 16/255, type = float, help = "Epsilon (strength) of RFGSM attack")
 
 
 args = parser.parse_args()
